[modify] orderBox 위젯 분리
